fix failing tests

+++ main/models.py
@@ -506,7 +506,7 @@ class Funding(models.Model):
             The total number of days of effort provided by the funding.
         """
         days_effort = round(self.budget / self.daily_rate, 1)
-        return days_effort
+        return float(days_effort)
 
     @property
     def funding_left(self) -> Decimal:
@@ -531,7 +531,7 @@ class Funding(models.Model):
         Returns:
             The number of days worth of effort left.
         """
-        return round(self.funding_left / self.daily_rate, 1)
+        return float(round(self.funding_left / self.daily_rate, 1))
 
     @property
     def monthly_pro_rata_charge(self) -> float | None:
